+++ spec/models/custom_content_spec.rb
@@ -20,6 +20,29 @@ require "spec_helper"
 describe CustomContent do
   subject { custom_contents(:login) }
 
+  context "scopes" do
+    let!(:layer_custom_content) { Fabricate(:custom_content, context: Group.root) }
+    let!(:invoice_config_custom_content) { Fabricate(:custom_content, context: invoice_configs(:top_layer)) }
+
+    it "does not return custom contents from layer groups per default" do
+      expect(CustomContent.all).to include(subject)
+      expect(CustomContent.all).not_to include(layer_custom_content)
+      expect(CustomContent.all).not_to include(invoice_config_custom_content)
+    end
+
+    it "in_context only returns custom contents from group context" do
+      expect(CustomContent.in_context(Group.root)).not_to include(subject)
+      expect(CustomContent.in_context(Group.root)).to include(layer_custom_content)
+      expect(CustomContent.in_context(Group.root)).not_to include(invoice_config_custom_content)
+    end
+
+    it "in_context only returns custom contents from invoice config context" do
+      expect(CustomContent.in_context(invoice_configs(:top_layer))).not_to include(subject)
+      expect(CustomContent.in_context(invoice_configs(:top_layer))).not_to include(layer_custom_content)
+      expect(CustomContent.in_context(invoice_configs(:top_layer))).to include(invoice_config_custom_content)
+    end
+  end
+
   context ".list" do
     it "contains one entry per main item" do
       expect(CustomContent.list.size).to eq(CustomContent.count)
@@ -38,6 +61,13 @@ describe CustomContent do
     it "creates list with several elements" do
       subject.placeholders_required = "login-url, foo ,bar"
       expect(subject.placeholders_required_list).to eq(%w[login-url foo bar])
+    end
+
+    it "creates list with several elements from main custom content" do
+      subject.update!(placeholders_required: nil)
+      context_custom_content = Fabricate(:custom_content, context: Group.root, key: subject.key)
+      subject.update!(placeholders_required: "login-url, foo ,bar", body: "{login-url}{foo}{bar}")
+      expect(context_custom_content.placeholders_required_list).to eq(%w[login-url foo bar])
     end
   end
 
@@ -66,6 +96,12 @@ describe CustomContent do
       subject.subject = "Mail from {sender}"
 
       is_expected.to be_valid
+    end
+
+    it "fails in context when required placeholder from main custom content is missing" do
+      context_custom_content = Fabricate(:custom_content, context: Group.root, key: subject.key, body: "{login-url}")
+      context_custom_content.body = ""
+      expect(context_custom_content).not_to be_valid
     end
   end